thread_monitor_impl_win32/linux split into platform-independent thread_monitor and platform-specific process_explorer utilities
- this change is required to support macos' thread_info for thread times.

+++ frontend/src/threads_model.cpp
@@ -40,7 +40,7 @@ namespace micro_profiler
 	{
 		const const_iterator i = find(thread_id);
 
-		return end() != i ? native_id = i->second.native_id, true : false;
+		return end() != i ? native_id = static_cast<unsigned int>(i->second.native_id), true : false;
 	}
 
 	bool threads_model::get_key(unsigned int &thread_id, index_type index) const throw()
